backend/yolo/requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
PyYAML>=5.3.1
tqdm>=4.64.0
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
scipy>=1.4.1
mediapipe>=0.10.0
flask>=2.0.0
supabase>=2.0.0
python-dotenv>=0.19.0
